TodoTaskPolicy Implemented

+++ app/Http/Controllers/TodoTaskController.php
@@ -14,6 +14,9 @@ class TodoTaskController extends Controller
     }
 
     public function update(TodoTask $todoTask, TodoTaskUpdateRequest $request){
+
+        $this->authorize('update',$todoTask);
+
         $input = $request->validated();
 
         $todoTask->fill($input);
@@ -23,6 +26,9 @@ class TodoTaskController extends Controller
     }
 
     public function destroy(TodoTask $todoTask){
+
+        $this->authorize('destroy',$todoTask);
+
         $todoTask->delete();
     }
 }
